Alteração da verificação da permissão nos controllers.

+++ app/Http/Controllers/CursoController.php
@@ -12,13 +12,10 @@ use App\Events\CrudEvent;
 use App\Http\Requests\CursoRequest;
 use App\Repositories\CursoRepository;
 use App\Repositories\RegionalRepository;
-use App\Traits\ControleAcesso;
 use Illuminate\Support\Facades\Request as IlluminateRequest;
 
 class CursoController extends Controller
 {
-    use ControleAcesso;
-
     private $class = 'CursoController';
     private $cursoModel;
     private $cursoRepository;
@@ -36,10 +33,10 @@ class CursoController extends Controller
 
     public function index()
     {
-        $this->autoriza($this->class, __FUNCTION__);
+        $this->authorize('viewAny', auth()->user());
         $resultados = $this->cursoRepository->getToTable();
         $tabela = $this->cursoModel->tabelaCompleta($resultados);
-        if(!$this->mostra($this->class, 'create'))
+        if(auth()->user()->cannot('create', auth()->user()))
             unset($this->variaveis['btn_criar']);
         $variaveis = (object) $this->variaveis;
         return view('admin.crud.home', compact('tabela', 'variaveis', 'resultados'));
@@ -47,7 +44,7 @@ class CursoController extends Controller
 
     public function create()
     {
-        $this->autoriza($this->class, __FUNCTION__);
+        $this->authorize('create', auth()->user());
         $variaveis = (object) $this->variaveis;
         $regionais = $this->regionalRepository->getRegionais();
         return view('admin.crud.criar', compact('variaveis', 'regionais'));
@@ -55,6 +52,8 @@ class CursoController extends Controller
 
     public function store(CursoRequest $request)
     {
+        $this->authorize('create', auth()->user());
+
         $request->validated();
         
         $save = $this->cursoRepository->store($request);
@@ -69,7 +68,7 @@ class CursoController extends Controller
 
     public function edit($id)
     {
-        $this->autoriza($this->class, __FUNCTION__);
+        $this->authorize('updateOther', auth()->user());
         $resultado = Curso::with('regional','user')->findOrFail($id);
         $regionais = $this->regionalRepository->getRegionais();
         $variaveis = (object) $this->variaveis;
@@ -78,6 +77,8 @@ class CursoController extends Controller
 
     public function update(CursoRequest $request, $id)
     {
+        $this->authorize('updateOther', auth()->user());
+
         $request->validated();
 
         $update = $this->cursoRepository->update($id, $request);
@@ -92,7 +93,7 @@ class CursoController extends Controller
 
     public function destroy($id)
     {
-        $this->autoriza($this->class, __FUNCTION__);
+        $this->authorize('delete', auth()->user());
         $curso = $this->cursoRepository->getById($id);
 
         $delete = $curso->delete();
@@ -107,7 +108,7 @@ class CursoController extends Controller
 
     public function lixeira()
     {
-        $this->autorizaStatic(['1']);
+        $this->authorize('onlyAdmin', auth()->user());
         $resultados = $this->cursoRepository->getTrashed();
         $variaveis = (object) $this->variaveis; 
         $tabela = $this->cursoModel->tabelaTrashed($resultados);
@@ -116,7 +117,7 @@ class CursoController extends Controller
 
     public function restore($id)
     {
-        $this->autorizaStatic(['1']);
+        $this->authorize('onlyAdmin', auth()->user());
         $curso = $this->cursoRepository->getTrashedById($id);
         
         $restore = $curso->restore();
@@ -131,36 +132,40 @@ class CursoController extends Controller
 
     public function inscritos($id)
     {
-        $this->autoriza('CursoInscritoController', 'index');
-        $resultados = CursoInscrito::where('idcurso', $id)
-            ->orderBy('created_at', 'desc')
-            ->paginate(10);
-        $curso = $this->cursoRepository->getById($id);
-        $now = date('Y-m-d H:i:s');
-        if(!$curso)
-            abort(500);
-        $variaveis = [
-            'pluraliza' => 'inscritos',
-            'plural' => 'inscritos',
-            'singular' => 'inscrito',
-            'singulariza' => 'o inscrito',
-            'continuacao_titulo' => 'em <strong>'.$curso->tipo.': '.$curso->tema.'</strong>',
-            'btn_lixeira' => '<a href="/admin/cursos" class="btn btn-default">Lista de Cursos</a>',
-            'busca' => 'cursos/inscritos/'.$id,
-            'addonsHome' => '<a href="/admin/cursos/inscritos/download/'.$id.'" class="btn btn-primary mb-2">Baixar CSV</a>'
-        ];
-        if($curso->datatermino >= $now) 
-            $variaveis['btn_criar'] = '<a href="/admin/cursos/adicionar-inscrito/'.$curso->idcurso.'" class="btn btn-primary mr-1">Adicionar inscrito</a> ';
-        if(!$this->mostra('CursoInscritoController', 'create'))
-            unset($variaveis['btn_criar']);
-        $tabela = CursoInscritoController::tabelaCompleta($resultados, $curso->idcurso);
-        $variaveis = (object) $variaveis;
-        return view('admin.crud.home', compact('tabela', 'variaveis', 'resultados'));
+        if(perfisPermitidos('CursoInscritoController', 'index'))
+        {
+            $resultados = CursoInscrito::where('idcurso', $id)
+                ->orderBy('created_at', 'desc')
+                ->paginate(10);
+            $curso = $this->cursoRepository->getById($id);
+            $now = date('Y-m-d H:i:s');
+            if(!$curso)
+                abort(500);
+            $variaveis = [
+                'pluraliza' => 'inscritos',
+                'plural' => 'inscritos',
+                'singular' => 'inscrito',
+                'singulariza' => 'o inscrito',
+                'continuacao_titulo' => 'em <strong>'.$curso->tipo.': '.$curso->tema.'</strong>',
+                'btn_lixeira' => '<a href="/admin/cursos" class="btn btn-default">Lista de Cursos</a>',
+                'busca' => 'cursos/inscritos/'.$id,
+                'addonsHome' => '<a href="/admin/cursos/inscritos/download/'.$id.'" class="btn btn-primary mb-2">Baixar CSV</a>'
+            ];
+            if($curso->datatermino >= $now) 
+                $variaveis['btn_criar'] = '<a href="/admin/cursos/adicionar-inscrito/'.$curso->idcurso.'" class="btn btn-primary mr-1">Adicionar inscrito</a> ';
+            if(auth()->user()->cannot('create', auth()->user()))
+                unset($variaveis['btn_criar']);
+            $tabela = CursoInscritoController::tabelaCompleta($resultados, $curso->idcurso);
+            $variaveis = (object) $variaveis;
+            return view('admin.crud.home', compact('tabela', 'variaveis', 'resultados'));
+        } else
+            abort(403);
+        
     }
 
     public function busca()
     {
-        $this->autoriza($this->class, 'index');
+        $this->authorize('viewAny', auth()->user());
         $busca = IlluminateRequest::input('q');
         $resultados = Curso::where('tipo','LIKE','%'.$busca.'%')
             ->orWhere('tema','LIKE','%'.$busca.'%')
